去除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@22800 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cat/catDCLogMgr.cpp
@@ -362,7 +362,7 @@ namespace engine
             break ;
          }
          /// max time check
-         if ( maxTime > 0 && time( NULL ) - bTime >= maxTime )
+         if ( maxTime > 0 && time( NULL ) - bTime >= (UINT64)maxTime )
          {
             break ;
          }
